Remove virtual definitions until we use them.
Besides, it's more likely that the concept of virtual definition is
more naturally tied to Forms. One possibility is to add a PName
to a (Form : Exp) that tells you which eliminator to use on that form.
For example, the default is "Form #Bin #elimBin []", but if you write
Nat = Bin@elimBinAsNat, then this translates to "Form #Bin #elimBinAsNat []".
So, instead of  having a "virtual Nat" definition, the underlying type is really
just an ordinary "Bin", except we eliminate it with #elimBinAsNat. As Francisco
mentioned, the names of the methods in #elimBinAsNat can be used as the names of
the patterns when doing pattern matching.

Forms with different eliminator PNames should not be distinguished by alpha and
beta (conversion) equality.