Change multiple construtor syntax to separate by slashes.
And change back multiple constructor arguments to be separated by space.

Multiple constructors cannot be separated by spaces because of the optional
colon at the end, so we don't know if the last argument is the type or the
name of a constructor. So, for multiple constructors we use a slash, which
looks pretty nice.

Searching for beauty.