Pretty print locally unique names.
Type checking is still done with globally unique names to ease reasoning.

Other changes:

The --verbose option prints the environment even when
the program is type correct. Before it would only print
the environment in error messages.

Primitive `#' sigils are not pretty printed.

Get rid of contexts, in favor of telescopes everywhere.
It is okay to store context telescopes in legible order,
rather than lookup order, because variable names are unique.