Make telescope refinement preserve dependency order.
Telescopes are refined when searching for a covering.
Unification of type indices results in a subtitution that
is used to refine the telescope. Because type arguments
can share indices, and prior pattern arguments can be
inaccessibles instead of bindings, refining a telescope
used to sometimes result in telescopic arguments whose
types reference variables that occur later in the telescope.

Because internally variables are unique, this order did not matter
most of the time. However, during pretty printing the ordering
bug is obvious.