Mark inaccessible variables when checking pattern matching functions.
Prior to this change, the user may have seen variables in goal
holes that they actually could not refer to in the RHS of
a pattern clause. Such variables are needed to type functions,
but the user might not bind them into scope. Now such variables
are prefixed by a "period" character.